requirements.txt
# 最小依赖版本
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-chroma>=0.1.0
openai>=1.6.0
chromadb>=0.4.0
python-dotenv>=1.0.0
language-tool-python>=3.2.0
